select from data source has been supported
New entities appeared in the ydb - external data sources. Before this review, it was impossible to read the data in the sql queries from these objects. Reading from these objects is supported in this review. An example of such a reading:
```(sql)
SELECT * FROM `/local/s3_data_source`.`/` WITH (
  format="json_each_row",
  schema(
    key Utf8 NOT NULL,
    value Utf8 NOT NULL
  )
)
```
Here `/local/s3_data_source` is an external data source that describes the connection to s3 storage