Proofread contrib/tools/swig licenses
`LicenseRef-scancode-swig` stands for

```
You may copy, modify, distribute, and make derivative works based on
this software, in source code or object code form, without
restriction. If you distribute the software to others, you may do
so according to the terms of your choice. This software is offered as
is, without warranty of any kind.
```