fix paramiko wrapper
If we do not explicitly close the file, an sftp connection, associated with it may stay open. This happens because `sftp` context manager does not limit the lifetime of the file object. If an sftp connection stays open, it will be impossible to execute the script in the next step, because the file will be locked.