fix deletion of rdma server session
Disconnected session was not deleted and all other sessions were deleted
instead due to problem in rcu list predicate implementation.

Just deleting the server session and calling DestroyQP is not enough and will
lead to deadlock. We need to wait for all WR's to flush and only then it's safe
to delete Connection object (calling rdma_destroy_id)