NEBDUTY-965: promise object passed as pointer to different thread released too early
promise object is passed as pointer to different thread which will invoke
SetValue on promise. The main thread which is waiting in
`promise.GetFuture().Wait()` will get control and delete the promise variable
keeping dangling pointer in the other thread.

Since the calling thread won't access the dangling pointer nothing bad should happen
but this triggers failure of unit tests compiled with tsan