NEBDUTY-965: Remove unsafe debug code in vhost-server aio backend
This code is unsafe because it attempts to print the contents of a batch after
it has been submitted. This can lead to use-after-free errors if the IOs in the
batch are freed before the print function is called.

Additionally, this code caused
`make --sanitize=address -tt cloud/blockstore/vhost-server/ut`
to fail due to use-after-free errors.