csi: refactor usage of volume ID as disk or filesystem ID
Now there is assumption that volume ID is always identical to NBS ID of
disk ar filesystem. This is not techically required and is about to
change in next commits.
So extract explicit assignment of diskId or filesystemId (nbsId in
generic code that is used by both FS and block devices).

Signed-off-by: Anton Kuchin <antonkuchin@nebius.com>