csi: support volumes that are unique per VM
Kubernetes treats PVs with the same volumeHanlde (that translates to CSI
VolumeId) as the same volume. So when a new pod starts using volume that
is already staged on the node then this volume is not staged again but
just published for a new pod.
This caused problems with kubevirt VMs:
- same filestore volume can be used by different VMs on a host, and each
  VM needs its own endpoint (so separate volume staging)
- when disk is disconnected from one VM and connected to other VM on the
  same host old endpoint shouldn't be reused

To resolve this problem we change the format of volumeHandle (VolumeId)
to include VM instanceId to guarantee different endpoints per VM, but
allow reusing endpoint for multiple pods of the same VM (for hotplug).

New VolumeId-s for kubevirt have format "nbsId#instanceId".
We keep supporting old VolumeId-s that consist just from "nbsId" to
allow migration.
When all disks are migrated new format should be enforced and
compatibility code can be removed.

Signed-off-by: Anton Kuchin <antonkuchin@nebius.com>