[Disk Manager] Create snapshot migration task to another database (#2820)
YDB storage does not support backing up tables with external blobs. So, the process of switching between databases (e.g. cluster migration) can't be performed right now. To avoid this issue, and provide seamless migration we've chosen semi-manual solution:

  * Migrate snapshots data manually one by one
  * Shut down controlplane for a short period of time for all the pending tasks to finish.
  * Migrate remaining snapshots
  * Clean up not required snapshots

Here is implementation of a task which migrates a single snapshot's data to a different database