[Filestore] Implement CPU throttling for cleanup (#3268)
Normally, Cleanup execution is scheduled in response to user operations that produce deletion markers.
Changing thresholds in the config may result in massive amount of deletion markers to be cleaned up.
This causes Cleanup to run incessantly for a long time (up to several hours) and it may result in performance degradation.

Solution: track the minimal amount of deletion markers since the last tablet reboot and throttle cleanup when the amount of deletion markers is expected to drop below the minimal amount.