[Disk Manager] Integration of cell selection mechanism into disk creation (#3638)
A prepareZoneId step has been added to the disk creation process. It calls the CellSelector to determine which cell should be used for disk creation. This functionality only works for non-local disks.

The CellSelector is initialized when the controlplane part of the service starts up. If its config is not specified, a default one is created.

Tests for the disks service have been modified. Test cases that verify different creation scenarios have been moved to common_test.go, while disk_service_test_cells.go contains cases where a disk is requested to be created in a zone divided into cells or in a specific cell within a zone.

Methods have been added to testcommon for proper test functionality.