[Disk Manager] Refactor InflightDuration and fix tests being flaky (#4061)
Part of #3738 

Refactor InflightDuration: move the logic from task ping to
`updateStateWithPreparation` function and update the tests accordingly.
Remove approximate checks for ModifiedAt and InflightDuration
(thresholds) because this caused the test to be flaky (see #4043).

<details>
<summary>Problems with previous implementation </summary>

Currently, InflightDuration updates at the moment of the ping. It uses
time.Since (which is equivalent to time.Now().Sub), and this leads to
accuracy problems: we lose a small amount of time (the ModifiedAt
timestamp is updated slightly later, and the delta between the
ModifiedAt update and the time.Since call is the time we fail to account
for).

This is impossible to test reliably, which we are now observing with
flapping tests related to InflightDuration.
</details>

Refactor runner_test.go as requested in
https://github.com/ydb-platform/nbs/pull/3760#discussion_r2171529982 and
https://github.com/ydb-platform/nbs/pull/3760#discussion_r2168760832.

~~Fix taskPinger in runner.go: if `ping()` returned and error and the
context was cancelled at the same time, the error would be suppressed,
even if it's not related to context cancel.~~ Rejected by svartmetal:
https://github.com/ydb-platform/nbs/pull/4061#discussion_r2281952127