[NBS, Filestore] Add "units" label to request counters (#4217)
**Summary**
This PR introduces a units label for request counters to ensure seamless
visualization in dashboards.

**Motivation**
The label will be required when transitioning histogram/percentile
metrics from milliseconds to microseconds, allowing graphs to remain
consistent across unit changes.

**Additional Changes**
A minor refactoring was performed as part of this update.
`TRequestPercentiles` should be consistent with THistBase in order to:
* inherit the same name as the source histogram
* inherit the same units as the source histogram
* update its values from the source histogram