Fix TStatsAggregator (#4230)
**Summary**

Update TStatsAggregator to be compatible with the recent changes
introduced in
[ydb-platform/nbs#4029](https://github.com/ydb-platform/nbs/pull/4029)

**Motivation**

Recent changes in [PR
#4029](https://github.com/ydb-platform/nbs/pull/4029) modified the way
counters are handled. As a result, TStatsAggregator was no longer
compatible, causing several load test to fail
(`cloud/blockstore/tests/loadtest/local*`). This PR aligns
TStatsAggregator with the new behavior and restores the affected tests.