[Blockstore] Add new parameter to corruption tests: path to file with list of compute nodes (#4076)
Add new parameter to corruption tests: path to the file listing compute
nodes. One of these nodes will be chosen to run the test on. We already
have '--compute-node' parameter, but now we add ability to pass the list
of nodes instead of single node.

To to this, we had to add the corresponding parameter to ycp client in
blockstore/pylibs.

If the new parameter is not specified, behaviour will not change.