[NBS] fix "disk.{read,write}_throttler_delay" user counter (#4254)
**Motivation**

The user counters `disk.read_throttler_delay` and
`disk.write_throttler_delay` are currently broken (after
https://github.com/ydb-platform/nbs/pull/4217). They rely on the
`units=usec` label, but the underlying `ThrottlerDelay` counter
(calculated via `volume_service`) does not provide this label. As a
result, these counters do not work as intended.

**Summary**

This PR updates the implementation of the user counters so that they
function correctly regardless of the presence or absence of the units
label. In addition, proper tests are introduced to cover this behavior
and prevent regressions.