issue-4138: StrictFileSystemSizeEnforcementEnabled flag is added (#4137)
References #4138

When `StrictFileSystemSizeEnforcementEnabled` is on, shards know about
other shards.
In this case, when the counters are updated
(TIndexTabletActor::HandleUpdateCounters)
`AggregateUsedBytesCount` and `AggregateUsedNodesCount` are the same in
all the shards
and actually values aggregated by all the shards and the main file
system.
These values will later be used for strict file system size enforcement.