TLoadFreshBlobsActor should read data from fresh channels starting from trimFreshLogToCommitId, not earlier (#4301)
It turned out that TLoadFreshBlobsActor can read blobs starting from
channel gen:step and not from trimFreshLogToCommitId. This leads to
extra reads of already useless data and could cause OOMs if TrimFreshLog
operation did not work for long time. Here trimFreshLogToCommitId is
used to restrict earliest blob commit id we are interested in, because
it is set during Flush and defines commit with no useful data below it.

This is a follow up for https://github.com/ydb-platform/nbs/pull/4266

I thought we could set barrier(trimFreshLogToCommitId) for fresh channel
during startup, but it turned out that EvRange after CollectGarbage can
read "phantom" blobs and also such CollectGarbage can set barriers in
the "past" which looks strange and useless