issue-4231 use per-directory IsExhaustive flag for stats of non-existent files (#4337)
#4231

Imagine a situation where we know that all children of a given node are
in memory, and a client tries to do stat(node, non-existent-children).
In those situations, we should not perform a real transaction, looking
up a non-existent row in a NodeRefs table, but we should use only
in-memory information.

This is what this PR does