[DM] Move estimated times from control-plane to data-plane (#4300)
Part of #3738

### 1. Move estimated times from control-plane tasks to data-plane tasks

Disks:
- `disks.CreateDiskFromImage` -> `dataplane.TransferFromSnapshotToDisk`
- `disks.CreateDiskFromSnapshot` ->
`dataplane.TransferFromSnapshotToDisk`
- `disks.MigrateDisk` - `dataplane.ReplicateDisk`

Images:
- `images.CreateImageFromDisk` -> `dataplane.CreateSnapshotFromDisk`
- `images.CreateImageFromImage` ->
`dataplane.CreateSnapshotFromSnapshot`,
`dataplane.CreateSnapshotFromLegacySnapshot`
- `images.CreateImageFromURL` -> `dataplane.CreateSnapshotFromURL`
- `images.CreateImageFromSnapshot` ->
`dataplane.CreateSnapshotFromSnapshot`,
`dataplane.CreateSnapshotFromLegacySnapshot`

Snapshots:
- `snapshots.CreateSnapshotFromDisk` ->
`dataplane.CreateSnapshotFromDisk`

### 2. Add estimates to dataplane tasks

- `dataplane.CreateDRBasedDiskCheckpoint`
- `dataplane.DeleteSnapshotData`
- `dataplane.TransferFromDiskToDisk`

Introduce bandwidth for `CreateDRBasedDiskCheckpoint` and set it to 200
MiB/s.

Introduce bandwidth for `TransferFromDiskToDisk` and set it to 300 MiB/s
(copy from `ReplicateDisk`).

Introduce bandwidth for `DeleteSnapshotData` and set it to 3000 MiB/s.
This task removes references to chunks (and possibly sends a request to
S3 to remove them if the reference count reaches zero), so it should
have the same speed as the shallow copy (creating references) in
`CreateSnapshotFromSnapshot`.

### 3. Refactor performance config

Drop all previous fields.

Add 4 general fields:
- `SnapshotShallowCopyBandwidthMiBs` = 3000
- `TransferFromDiskToDiskBandwidthMiBs` = 200
- `TransferBetweenDiskAndSnapshotBandwidthMiBs` = 200
- `TransferFromURLToSnapshotBandwidthMiBs` = 200

Add 1 field for specific dataplane tasks:

- `CreateDRBasedDiskCheckpointBandwidthMiBs` = 300

**Note: if you have one of performance config fields changed in your
config fields, you will have to update them to use new names**.

### 4. Enhance estimates for CreateSnapshotFromDisk

Previously, estimates for the data-plane task were set in the
corresponding control-plane task (`snapshots.CreateSnapshotFromDisk`).
This control-plane task scheduled the data-plane task, waited for it to
finish, and calculated an estimate based on the transferred chunk count.

This is incorrect: the estimate should be set before the actual data
transfer (or other long-running actions) occurs, so the hanging tasks
logic can know the ETA of this task.

But `CreateSnapshotFromDisk` can perform two types of copies: regular
and (for incremental snapshots) shallow. A regular copy physically
transfers data from the source to the target with a bandwidth of 200
MiB/s. For a shallow copy, however, it only transfers data for changed
chunks, and the other chunks are just referenced.

Setting the estimate for this task as `Estimate(diskSize, 200 MiB/s)`
would be a huge overestimation for incremental snapshot cases where only
a small amount of data has changed.

I am proposing to use the following logic for disks with incremental
snapshots: calculate `transferDuration = Estimate(bytesToTransfer, 200
MiB/s)` and `shallowCopyDuration = Estimate(snapshotSize, 3000 MiB/s)`,
and use `max(transferDuration, shallowCopyDuration)` as the final
estimate (data transfer and shallow copy are performed in parallel). For
regular snapshots, previous logic will apply.