switch to deadlines in collect garbage request (#4374)
This is a follow up for https://github.com/ydb-platform/nbs/pull/4266

We finally decided to support and use the deadline parameter in
TEvBlobStorage::EvCollectGarbage.

According to the YDB team, this functionality has been available since
24-1, and in 25-1 tests for deadlines in the BS proxy were implemented.
These tests are green, and since the functionality has not changed since
24-1, we may assume that it also works in 24-1.

Deadlines are now supported in both CollectGarbage and TrimFreshLog
operations.