[Disk Manager] Move disk zone definition inside tasks.  (#4023)
https://github.com/ydb-platform/nbs/issues/4132

Move disk zone (cell) definition inside task for future cell selection
with database access

Moving the zone (cell) definition for the future disk inside the task,
as we will need to query the database for shard selection later. This
way, if we encounter a retriable error, we'll be able to retry the
operation properly.


We no longer need to check whether a disk with this id has already been
created in the specified zone, since the saving selected zone inside
task state.