[DM] Add estimates for erasing local disks in disks.DeleteDisk (#4390)
Part of #3738 

Add estimated times for `disks.DeleteDisk`. In most cases, the task
completes instantly. But on deleting local disks (local-hdd or
local-ssd) blockstore needs to erase it, so it takes some time that I am
estimating in this PR.

Proposing setting the bandwidth for SDD local disks to 500 MiB/s and HDD
local disks to 100 MiB/s.