[DM] Allow overriding EstimatedInflightDuration and EstimatedStallingDuration via tasks config (#4432)
Part of #3738 

Add `EstimatedInflightDurationOverrideByTaskType` and
`EstimatedStallingDurationOverrideByTaskType` fields to config which
handles keys as task types and values as estimated durations.

If set, estimates for selected task types will be automatically set in
`lockTaskToExecute`. As SetEstimatedInflightDuration and
SetEstimatedStallingDuration allows to set estimate only once, this
estimate won't be overrided by the task.

This allows temporarily changing estimates for tasks without needing to
change the code and deploying new release.