add parameter to override target when reading additional files for seed target (#4453)
In most cases seed config is the same as for nbs-control, but sometimes
this is not true and we have to create separate target for seed. However
in this case we need to have separate folders for files like
nbs-domains.txt and nbs-names.txt.

This pr allows to specify folder to take these files from (in most cases
nbs-domains.txt and nbs-names.txt are ok to be taken from nbs-control
folder)