issue-4138: Strictly enforce file system size (#4372)
When `StrictFileSystemSizeEnforcementEnabled` is on, shards know about
other shards.
In this case, when the counters are updated
(TIndexTabletActor::HandleUpdateCounters)
`AggregateUsedBytesCount` and `AggregateUsedNodesCount` are the same in
all the shards and actual values aggregated by all the shards and the
main file system.
In addition, when 'Strict' is on sizes of all the shards are equal to
that of the main file system.
So, it is possible now to correctly enforce the size limit in every
shard.


References #4138