Introduce TBlockStoreMethod template (#4554)
1. The main idea is to make for each type of requests enumerated by the
BLOCKSTORE_SERVICE macro (NProto::TReadBlocksRequest,
NProto::TCreateCheckPointRequest, etc.) your template
TBlockStoreMethod<TRequest, TResponse>
2. In this template TBlockStoreMethod, there will be convenient access
to the types of request and response, the name of the request, and some
query classifications, for example, IsReadWriteRequest()
3. Such a template TBlockStoreMethod is easy to get by request type or
response type.
4. TBlockStoreMethod allows you to call the desired method from the
IBlockStore
5. TBlockStoreImpl does the opposite transformation, inherits from
IBlockStore, and when calling the virtual method, IBlockStore dispatches
the call to the template method of the user class, immediately passing
the TBlockStoreMethod type to it in the template. Which allows you to
write IBlockStore heirs without macros.