Add TAdaptiveTimeHist that allows to collect either ms or us metrics (#4610)
The PR https://github.com/ydb-platform/nbs/pull/4029 has changed at
least two things:
1) Latency histograms for the server are reported in microseconds
2) Latency percentiles for server_volume are also reported in
microseconds

The first change was not necessary, but it was done to increase the
resolution of the server_volume percentiles.

My PR introduces a toggle that allows to report milliseconds for
histograms in the server, but still reports percentiles in microseconds
for server_volume.