issue-4138: Shard channel count reculacation during resize added (#4706)
If there is a filesystem which has a shard that has TotalBytesCount
larger than that of main FS,
during transition to StrictFileSystemSizeEnforcementEnabled we calculate
less channels number for this shard
than it really has, and it leads to and error. 
It's hard (but theoretically possible) to have this kind of filesystem
in real life but we have one in our test environment.
So, recalculation of channel count that takes into account real number
of channels in a shard is added.