issue-2566: [Filestore] Don't use 7th byte in handles and node ids (#4736)
As we are going to use the 7th byte to store shard ids, we first need to
stop using it.
The 7th byte is now used by handles, as they are random numbers.
We first stop using the seventh and wait till handles that use it are
gone.
Newly added sensor "SevenBytesHandlesCount" will tell us when old
handles are gone.

#2566