issue-4794: [Disk Manager] Fix migration for incremental snapshots (#4775)
#4794
For incremental snapshots, database migration did not preserve the
latest snapshots for a disk and did not preserve the order of
incremental snapshots. Preserve the order of base snapshot ids. (Note
that snapshot data for incremental snapshots is still being duplicated,
but now there will be no problem with creating incremental snapshots
from the current one on a migration target db). This fix will allow to
perform a transfer with a shallow copy in the future.