Backport https://github.com/ydb-platform/ydb/pull/30354 (#4822)
issue #4823

This change was already done in ydb repo:
https://github.com/ydb-platform/ydb/pull/30354. Now we do it in nbs
repo.

Description:

We (NBS) are going to enable the discard feature for NBS disks. To
deploy it safely, we need to be able to turn on this feature at the
volume level. To do this, we need to add the corresponding flag to the
VolumeConfig.