[NBS] Fixed crash in TStatsServiceActor, which occurs because of a finished resync during snapshot taking (#4840)
Assume that volume is registered in TStatsServiceActor. If resync is
completed during snapshot taking, we will restart the shadow disk actor
and send a TEvStatsService::TEvRegisterVolume message to the
TStatsServiceActor. The TStatsServiceActor will replace the volume's
PerfCounters but will not register any new counters. As a result, the
next time we publish the volume counters, it will be considered to be
registered, and we will receive a crash during publication.