issue-1751: [Filestore] Restore and drain WriteBackCache at session recreation (#4711)
https://github.com/ydb-platform/nbs/issues/1751

Resolve data loss in the following scenario:
1. WriteBackCache was enabled and used.
2. ServerWriteBackCacheEnabled flag was disabled in the configuration.
3. nfs-vhost was restarted.

Actual behavior: nfs-vhost completely ignores the non-flushed requests
remaining in the persistent queue.

Expected behavior: nfs-vhost should check for the presence of the
persistent queue and initialize WriteBackCache if the persistent queue
exists regardless of the value of ServerWriteBackCacheEnabled flag. If
ServerWriteBackCacheEnabled is not set, new writes should not be
processed until the queue is flushed.