issue-4823: add EnableVhostDiscardForNewVolumes feature; enable Discards for volumes with VhostDiscardEnabled flag set in SchemeShard (#4811)
https://github.com/ydb-platform/nbs/issues/4823

Add EnableVhostDiscardForNewVolumes parameter to Storage config together
with the corresponding feature. The feature can be enabled for specific
clouds/folders, as well as for a percentage of clouds/folders.

- If the feature is enabled, then newly creared volumes will be created
with the VhostDiscardEnabled flag set in the VolumeConfig in
SchemeShard. (This flag was added here:
https://github.com/ydb-platform/nbs/pull/4822)
- If the volume has VhostDiscardEnabled set in SchemeShard, then
discards will be enabled for vhost endpoints for this volume (regardless
of VhostDiscardEnabled parameter in Server config).

Note that if VhostDiscardEnabled flag was once set in SchemeShard, it
should never be unset anymore for this volume.

Default behaviour does not change.