issue-4664: copy only origin blobs range (#4566)
issue: #4664

This refactoring(prefactoring) is required to remove unnecessary buffer
copy in TReadDataActor for zero-copy read
https://github.com/ydb-platform/nbs/blob/main/cloud/filestore/libs/storage/service/service_actor_readdata.cpp#L674
as it's not possible to copy data from aligned super range directly to
iovecs with unaligned origin range.

BufferOffset is always 0 in the response.



[Here](https://github.com/ydb-platform/nbs/pull/4790/commits/1bc7e40f6a134b91934fc6fd46fb41a8660e20df#diff-24b44c319c6982823d902343c7ea19cc99303e62f2ace02206de81dceaf7f6e9R152)
is the pr with elimination of buffer copy for iovecs over the current
branch